Se creó archivo cli para ejecutar la función mdlinks y se creó modulo instalable close #10
